Lifting out microrpc.
It probably makes sense to try to replace this with [tinyrpc].
